requirements.txt

django==4.1
django-environ==0.9.0
pillow==9.2.0
stripe==4.0.2
django-allauth==0.51.0
django-tailwind==3.4.0
django-crispy-forms==1.14.0
crispy-tailwind==0.5.0
psycopg2==2.9.3
djangorestframework==3.13.1